Role

+++ app/Http/Controllers/Backend/DepartmentBackendController.php
@@ -15,29 +15,34 @@ use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Response;
+use Symfony\Component\HttpFoundation\Response as SymfonyHttp;
 use function Pest\Laravel\delete;
 
 class DepartmentBackendController extends Controller
 {
     public function __construct(
-        public DepartmentRepositoryInterface $repository,
-        public SweetAlertFactory $factory
+        protected readonly DepartmentRepositoryInterface $repository,
+        protected readonly SweetAlertFactory $factory
     ) {
     }
 
     public function index(): Renderable
     {
-        return view('backend.domain.academic.departments.index', [
-            'departments' => $this->repository->getDepartments(),
-        ]);
+
+        $departments = $this->repository->getDepartments();
+
+        $this->authorize('Department-list', $departments);
+
+        return view('backend.domain.academic.departments.index', compact('departments'));
     }
 
     public function show(string $key): Factory|View|Application
     {
-        return view('backend.domain.academic.campus.departments.show', [
-            'department' => $this->repository->showDepartment(key:  $key),
-        ]);
+        $department = $this->repository->showDepartment(key:  $key);
+
+        return view('backend.domain.academic.departments.show', compact('department'));
     }
 
     public function create(): Renderable
@@ -47,6 +52,8 @@ class DepartmentBackendController extends Controller
 
     public function store(DepartmentRequest $attributes): RedirectResponse
     {
+        abort_if(Gate::denies('Department-create'), SymfonyHttp::HTTP_FORBIDDEN, '403 Forbidden');
+
         $this->repository->stored(attributes: $attributes, factory: $this->factory);
 
         return to_route('admins.academic.departments.index');
@@ -54,13 +61,15 @@ class DepartmentBackendController extends Controller
 
     public function edit(string $key): Factory|View|Application
     {
-        return view('backend.domain.academic.departments.edit', [
-            'department' => $this->repository->showDepartment(key:  $key),
-        ]);
+        $department = $this->repository->showDepartment(key:  $key);
+
+        return view('backend.domain.academic.departments.edit', compact('department'));
     }
 
     public function update(string $key, DepartmentRequest $attributes): RedirectResponse
     {
+        abort_if(Gate::denies('Department-edit'), SymfonyHttp::HTTP_FORBIDDEN, '403 Forbidden');
+
         $this->repository->updated(key: $key, attributes: $attributes, factory: $this->factory);
 
         return Response::redirectToRoute('admins.academic.departments.index');
@@ -68,6 +77,8 @@ class DepartmentBackendController extends Controller
 
     public function destroy(string $key): RedirectResponse
     {
+        abort_if(Gate::denies('Department-delete'), SymfonyHttp::HTTP_FORBIDDEN, '403 Forbidden');
+
         $this->repository->deleted(key: $key, factory: $this->factory);
 
         return Response::redirectToRoute('admins.academic.departments.index');
